tests/wpt/web-platform-tests/websockets/constructor/006.html

<!doctype html>
<title>WebSockets: converting first arguments</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
async_test(function(t) {
  var a = document.createElement('a');
  a.href = SCHEME_DOMAIN_PORT+'/echo';
  var ws = new WebSocket(a); // should stringify arguments; <a> stringifies to its .href
  assert_equals(ws.url, a.href);
  ws.onopen = t.step_func(function(e) {
    ws.send('test');
  });
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, 'test');
    ws.onclose = t.step_func(function(e) {
      ws.onclose = t.unreached_func();
      t.step_timeout(() => t.done(), 50);
    });
    ws.close();
  });
  ws.onerror = ws.onclose = t.unreached_func();
});
</script>
